Moves user context into a context directory
