+++ src/components/sign-in.tsx
@@ -1,6 +1,6 @@
 import { useAuthUrl } from 'hooks/use-auth-url';
 import { useSignIn } from 'hooks/use-sign-in';
-import React, { createContext, ReactNode, useState } from 'react';
+import React from 'react';
 import { Scope } from 'types/scope';
 import { FractalUser } from 'types/user';
 
@@ -15,30 +15,6 @@ export interface SignInProps {
    * See src/types/scope.ts for a list of available scopes.
    */
   scopes?: Scope[];
-}
-
-interface UserContextState {
-  setUser: (user: FractalUser | undefined) => void;
-  user?: FractalUser;
-}
-
-export const UserContext = createContext<UserContextState>({
-  setUser: () => undefined,
-  user: undefined,
-});
-
-export interface UserContextProviderProps {
-  children: ReactNode;
-}
-
-export function UserContextProvider({ children }: UserContextProviderProps) {
-  const [user, setUser] = useState<FractalUser | undefined>(undefined);
-
-  return (
-    <UserContext.Provider value={{ setUser, user }}>
-      {children}
-    </UserContext.Provider>
-  );
 }
 
 export function SignIn({ clientId, onError, onSuccess, scopes }: SignInProps) {
